Code reorganization, no changes to functionality
