+++ client/adventureMap/CTerrainRect.cpp
@@ -10,19 +10,21 @@
 #include "StdInc.h"
 #include "CTerrainRect.h"
 
-#include "mapHandler.h"
-#include "MapView.h"
 #include "CAdvMapInt.h"
 
 #include "../CGameInfo.h"
+#include "../CMT.h"
 #include "../CPlayerInterface.h"
-#include "../gui/CursorHandler.h"
 #include "../gui/CGuiHandler.h"
+#include "../gui/CursorHandler.h"
+#include "../mapRenderer/mapHandler.h"
 #include "../render/CAnimation.h"
 #include "../render/IImage.h"
 #include "../renderSDL/SDL_Extensions.h"
 #include "../widgets/TextControls.h"
-#include "../CMT.h"
+#include "../mapRenderer/MapView.h"
+#include "../mapRenderer/MapViewController.h"
+#include "../mapRenderer/MapViewModel.h"
 
 #include "../../CCallback.h"
 #include "../../lib/CConfigHandler.h"
